+++ src/JTypeInstruction.java
@@ -8,9 +8,9 @@ public class JTypeInstruction implements Instruction{
 //    mnemonic = "j"
 //    opcode = (hex >> 26) OR opcode = bx000010 because it's always the same
 //    index = (hex >> 0) & bx 11 1111 1111 1111 1111 1111 1111
-    public JTypeInstruction(int hex) {
+    public JTypeInstruction(long hex) {
         opcode = 0x2;
-        index = (hex >> 0) & 0x3FFFFFF;
+        index = (int)(hex >> 0) & 0x3FFFFFF;
         mnemonic = "j";
     }
